// src/Model/HeroModel.ts
export interface HeroModel
{
    id: number;
    image: string;
    imageDefault: string;
    imageDie: string;
    perkDescription: string;
    name: string;
    age: number;
    power: number;
    agility: number;
    intelligence: number;
    charisma: number;
    healthPoints: number;
    maxHealthPoints: number
    dodge: number;
    energy: number;
    maxEnergy: number;
    status: boolean;
}